% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.array.interval}
\alias{as.array.interval}
\title{Convert Interval to Array}
\usage{
\method{as.array}{interval}(x, ...)
}
\arguments{
\item{x}{An interval object.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
An array representation of the interval.
}
\description{
Convert Interval to Array
}
\examples{
as.array(inter_city)
}
