% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.data.frame.interval}
\alias{as.data.frame.interval}
\title{Convert Interval to Data Frame}
\usage{
\method{as.data.frame}{interval}(x, ...)
}
\arguments{
\item{x}{An interval object.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
A data frame representation of the interval.
}
\description{
Convert Interval to Data Frame
}
\examples{
as.data.frame(inter_city)
}
