% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.interval.array}
\alias{as.interval.array}
\title{Convert Array to Interval}
\usage{
\method{as.interval}{array}(x)
}
\arguments{
\item{x}{An array object.}
}
\value{
An interval object constructed from the array or attempts conversion.
}
\description{
Convert Array to Interval
}
