% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.interval.default}
\alias{as.interval.default}
\title{Default Method for Interval Conversion}
\usage{
\method{as.interval}{default}(x)
}
\arguments{
\item{x}{An object to be converted.}
}
\value{
`NULL` as default behavior for non-supported types.
}
\description{
Default Method for Interval Conversion
}
