% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.interval.interval}
\alias{as.interval.interval}
\title{Identity Conversion for Interval}
\usage{
\method{as.interval}{interval}(x)
}
\arguments{
\item{x}{An interval object.}
}
\value{
The interval object itself.
}
\description{
Identity Conversion for Interval
}
