% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.interval.matrix}
\alias{as.interval.matrix}
\title{Convert Matrix to Interval}
\usage{
\method{as.interval}{matrix}(x)
}
\arguments{
\item{x}{A matrix object.}
}
\value{
An interval object constructed from the matrix.
}
\description{
Convert Matrix to Interval
}
\examples{
as.interval(matrix(1:12, 3, 4))
}
