% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.interval.numeric}
\alias{as.interval.numeric}
\title{Convert Numeric to Interval}
\usage{
\method{as.interval}{numeric}(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
An interval object constructed from the numeric vector.
}
\description{
Convert Numeric to Interval
}
\examples{
as.interval(1:6)
}
