% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.matrix.interval}
\alias{as.matrix.interval}
\title{Convert Interval to Matrix}
\usage{
\method{as.matrix}{interval}(x, ...)
}
\arguments{
\item{x}{An interval object.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
A matrix representation of the interval.
}
\description{
Convert Interval to Matrix
}
\examples{
as.matrix(inter_city)
}
