% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.vector.interval}
\alias{as.vector.interval}
\title{Convert Interval to Vector}
\usage{
\method{as.vector}{interval}(x, ...)
}
\arguments{
\item{x}{An interval object.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
A vector representation of the interval.
}
\description{
Convert Interval to Vector
}
\examples{
as.vector(inter_city)
}
