% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REXPORT.R
\name{check_covering}
\alias{check_covering}
\title{Check a covering for a set of moduli and residuals}
\usage{
check_covering(r, m)
}
\arguments{
\item{r}{an integer vector containing residuals}

\item{m}{an integer vector containing the moduli}
}
\value{
a single boolean value
}
\description{
Checks whether or not a set of residuals and moduli.  Both parameters 
(r and m) must be integer vectors; use \code{\link{as.integer}} to convert 
generic numeric vectors to integer vectors.
}
