% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cluster_color}
\alias{cluster_color}
\title{Clustering Color for Plotting}
\usage{
cluster_color(x)
}
\arguments{
\item{x}{A clustering vector or matrix.}
}
\value{
A vector of colors corresponding to the clustering assignments.
}
\description{
Generate colors for data points based on their clustering assignments.
}
\examples{
plot(iris[, 1:2], col = cluster_color(neokm(iris, 2, 0.2, 0.05)$cluster))
}
