% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{covering}
\alias{covering}
\alias{as.cover}
\title{Cover class}
\usage{
covering(a = NA, b = NA, r = NA, p = NA, s = NA, h = NA,
  solution = list(k = NA, mod = NA), ...)

as.cover(x)
}
\arguments{
\item{a}{numeric vector containing remainders for the covering}

\item{b}{numeric vector containing moduli for the covering}

\item{r}{}

\item{p}{}

\item{s}{}

\item{h}{}

\item{solution}{}

\item{...}{}
}
\value{
a covering object
}
\description{
Coverings are list like objects consumed and produced by many of the
functions in this package.
}
\details{
The lengths of a, b must be the same.
}
\section{Functions}{
\itemize{
\item \code{as.cover}: Convert a list-like object to a covering
}}

