% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{degree2logical}
\alias{degree2logical}
\title{Convert Degree Matrix to Logical}
\usage{
degree2logical(x, t = min(apply(x, 1, max)))
}
\arguments{
\item{x}{A matrix of membership degrees.}

\item{t}{Threshold value for conversion.}
}
\description{
Transform a degree matrix into a logical matrix based on a threshold.
}
