% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_icmeans.R
\name{fuzzy_icmeans}
\alias{fuzzy_icmeans}
\title{Interval Fuzzy C-Means Clustering}
\usage{
fuzzy_icmeans(
  x,
  centers,
  m = 2,
  nstart = 2,
  distance = "euclid",
  trace = FALSE,
  iter.max = 40
)
}
\arguments{
\item{x}{A 3D interval array.}

\item{centers}{Either the number of clusters or pre-initialized centers.}

\item{m}{A number greater than 1 giving the degree of fuzzification.}

\item{nstart}{Number of executions to find the best result.}

\item{distance}{A string specifying the distance ('euclid', 'hausdorff').}

\item{trace}{Logical, if TRUE, tracing information on the progress of the
algorithm is produced.}

\item{iter.max}{Maximum number of iterations allowed.}
}
\value{
A list representing the clustering results.
}
\description{
Cluster interval data using the fuzzy c-means algorithm.
}
\examples{
fuzzy_icmeans(iaggregate(iris, col = 5), 2)
fuzzy_icmeans(iaggregate(iris, col = 5), iaggregate(iris, col = 5))
}
