% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{iaggregate}
\alias{iaggregate}
\title{Aggregate Data to 3D Interval Array}
\usage{
iaggregate(data, col = 1)
}
\arguments{
\item{data}{The data to aggregate.}

\item{col}{The column index to aggregate on.}
}
\description{
Aggregate data into a 3D interval array.
}
\examples{
iaggregate(iris, col = 5)
iaggregate(rock, col = 4)
iaggregate(cars, col = 1)
}
