% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{ibind}
\alias{ibind}
\title{Bind Multiple Interval Objects}
\usage{
ibind(..., class = FALSE)
}
\arguments{
\item{...}{Interval objects to bind.}

\item{class}{Logical indicating whether to create a new class by binding.}
}
\description{
Bind multiple interval objects into one.
}
\examples{
ibind(iaggregate(iris, 5), iaggregate(iris, 5))
ibind(iaggregate(iris, 5), iaggregate(iris, 5), iaggregate(iris, 5),
class = TRUE)
}
