% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ikmeans.R
\name{ikmeans}
\alias{ikmeans}
\title{Interval K-Means Clustering}
\usage{
ikmeans(
  x,
  centers,
  nstart = 10,
  distance = "euclid",
  trace = FALSE,
  iter.max = 20
)
}
\arguments{
\item{x}{A 3D interval array.}

\item{centers}{Either the number of clusters or pre-initialized centers.}

\item{nstart}{Number of executions to find the best result.}

\item{distance}{Distance metric ('euclid' or 'hausdorff').}

\item{trace}{Logical, if TRUE, trace progress of the algorithm.}

\item{iter.max}{Maximum number of iterations allowed.}
}
\value{
A list representing the clustering results.
}
\description{
Cluster interval data using the k-means algorithm.
}
\examples{
ikmeans(iaggregate(iris, col = 5), 2)
ikmeans(iaggregate(iris, col = 5), iaggregate(iris, col = 5))
}
