% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineokm.R
\name{ineokm}
\alias{ineokm}
\title{Interval Neo-KM Clustering}
\usage{
ineokm(
  x,
  centers,
  alpha = 0.3,
  beta = 0.05,
  nstart = 10,
  trace = FALSE,
  iter.max = 20
)
}
\arguments{
\item{x}{A 3D interval array.}

\item{centers}{Either the number of clusters or pre-initialized centers.}

\item{alpha}{A numeric value representing overlap.}

\item{beta}{A numeric value representing non-exhaustiveness.}

\item{nstart}{Number of executions to find the best result.}

\item{trace}{Logical, if TRUE, trace progress of the algorithm.}

\item{iter.max}{Maximum number of iterations allowed.}
}
\value{
A list representing the clustering results.
}
\description{
Cluster interval data using the Neo-KM algorithm.
}
\examples{
ineokm(iaggregate(iris, col = 5), 3)
ineokm(iaggregate(iris, col = 5), iaggregate(iris, col = 5), 1, 2)
}
