% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inter_city}
\alias{inter_city}
\title{Temperature by month and humidity in european city.}
\format{
A interval structure with 68 rows and 13 variables divided in min
and max with 17 class:
\describe{
  \item{temp.jan}{}
  \item{temp.fev}{}
  \item{temp.mars}{}
  \item{temp.avr}{}
  \item{temp.mai}{}
  \item{temp.juin}{}
  \item{temp.juil}{}
  \item{temp.aout}{}
  \item{temp.sep}{}
  \item{temp.oct}{}
  \item{temp.nov}{}
  \item{temp.dec}{}
  \item{humid}{}
}

Class :
\describe{
  \item{Allemagne}{}
  \item{Angleterre}{}
  \item{Autriche}{}
  \item{Belgique}{}
  \item{Bulgarie}{}
  \item{Croatie}{}
  \item{Danemark}{}
  \item{Espagne}{}
  \item{France}{}
  \item{Italie}{}
  \item{Pays-Bas}{}
  \item{Pologne}{}
  \item{Portugal}{}
  \item{Roumanie}{}
  \item{Russie}{}
  \item{Turquie}{}
  \item{Ukraine}{}
}
}
\usage{
inter_city
}
\description{
Temperature by month and humidity in european city.
}
\keyword{datasets}
