% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inter_emotions}
\alias{inter_emotions}
\title{Emotions in music aggregate on BPM to interval multi label data.}
\format{A interval structure with 59 rows and 71 variables divided in min
and max with 6 class:
\describe{
  \item{Mean_Acc1298_Mean_Mem40_Centroid}{}
  \item{Mean_Acc1298_Mean_Mem40_Rolloff}{}
  \item{Mean_Acc1298_Mean_Mem40_Flux}{}
  \item{Mean_Acc1298_Mean_Mem40_MFCC_0}{}
  ...
  \item{Mean_Acc1298_Mean_Mem40_MFCC_12}{}
  \item{Mean_Acc1298_Std_Mem40_Centroid}{}
  \item{Mean_Acc1298_Std_Mem40_Rolloff}{}
  \item{Mean_Acc1298_Std_Mem40_Flux}{}
  \item{Mean_Acc1298_Std_Mem40_MFCC_0}{}
  ...
  \item{Mean_Acc1298_Std_Mem40_MFCC_12}{}
  \item{Std_Acc1298_Mean_Mem40_Centroid}{}
  \item{Std_Acc1298_Mean_Mem40_Rolloff}{}
  \item{Std_Acc1298_Mean_Mem40_Flux}{}
  \item{Std_Acc1298_Mean_Mem40_MFCC_0}{}
  ...
  \item{Std_Acc1298_Mean_Mem40_MFCC_12}{}
  \item{Std_Acc1298_Std_Mem40_Centroid}{}
  \item{Std_Acc1298_Std_Mem40_Rolloff}{}
  \item{Std_Acc1298_Std_Mem40_Flux}{}
  \item{Std_Acc1298_Std_Mem40_MFCC_0}{}
  ...
  \item{Std_Acc1298_Std_Mem40_MFCC_12}{}
  \item{BH_LowPeakAmp}{}
  \item{BH_LowPeakBPM}{}
  \item{BH_HighPeakAmp}{}
  \item{BH_HighLowRatio}{}
  \item{BHSUM1}{}
  \item{BHSUM2}{}
  \item{BHSUM3}{}
}

Class :
\describe{
  \item{amazed.suprised}{}
  \item{happy.pleased}{}
  \item{relaxing.calm}{}
  \item{quiet.still}{}
  \item{sad.lonely}{}
  \item{angry.aggresive}{}
}}
\source{
\url{http://mulan.sourceforge.net/datasets-mlc.html}
}
\usage{
inter_emotions
}
\description{
Emotions in music aggregate on BPM to interval multi label data.
}
\keyword{datasets}
