% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inter_wine}
\alias{inter_wine}
\title{Results of a chemical analysis of wines grown in the same region in Italy but
derived from three different cultivars, aggregate on sulfur dioxide to
interval simple label data.}
\format{
A interval structure with 132 rows and 10 variables divided in min
and max with 7 class:
\describe{
  \item{fixed.acidity}{}
  \item{volatile.acidity}{}
  \item{citric.acid}{}
  \item{residual.sugar}{}
  \item{chlorides}{}
  \item{total.sulfur.dioxide}{}
  \item{density}{}
  \item{pH}{}
  \item{sulphates}{}
  \item{alcohol}{}
}

Class :
\describe{
  \item{Class3}{}
  \item{Class4}{}
  \item{Class5}{}
  \item{Class6}{}
  \item{Class7}{}
  \item{Class8}{}
  \item{Class9}{}
}
}
\source{
\url{https://archive.ics.uci.edu/dataset/186/wine+quality}
}
\usage{
inter_wine
}
\description{
Results of a chemical analysis of wines grown in the same region in Italy but
derived from three different cultivars, aggregate on sulfur dioxide to
interval simple label data.
}
\keyword{datasets}
