% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iokm.R
\name{iokm}
\alias{iokm}
\title{Interval OKM Clustering}
\usage{
iokm(
  x,
  centers,
  nstart = 10,
  distance = "euclid",
  algorithm = "std",
  update = "mean",
  trace = FALSE,
  iter.max = 20,
  secure = FALSE
)
}
\arguments{
\item{x}{A 3D interval array.}

\item{centers}{Either the number of clusters or pre-initialized centers.}

\item{nstart}{Number of executions to find the best result.}

\item{distance}{Distance measure ('euclid', 'hausdorff').}

\item{algorithm}{Algorithm type ('std': Standard, 'matrix': Matrix).}

\item{update}{Update method ('mean', 'sum', 'join', 'meet').}

\item{trace}{Logical, if TRUE, trace progress of the algorithm.}

\item{iter.max}{Maximum number of iterations allowed.}

\item{secure}{Logical, if TRUE, ensures min <= max in intervals.}
}
\value{
A list representing the clustering results.
}
\description{
Cluster interval data using the OKM algorithm.
}
\examples{
iokm(iaggregate(iris, col = 5), 2)
iokm(iaggregate(iris, col = 5), iaggregate(iris, col = 5))
}
