% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{is.interval}
\alias{is.interval}
\title{Interval Data Test}
\usage{
is.interval(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
Test if an object is a (strict) interval.
}
\examples{
is.interval(inter_city)
}
