% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{measure}
\alias{measure}
\title{Measure Overlap by Threshold}
\usage{
measure(x, min = 0, max = 1, step = 0.1)
}
\arguments{
\item{x}{A matrix of membership degrees.}

\item{min}{Minimum threshold value.}

\item{max}{Maximum threshold value.}

\item{step}{Step size for threshold values.}
}
\description{
Plot overlap as a function of threshold for membership degrees.
}
