% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neokm.R
\name{neokm}
\alias{neokm}
\title{NEOKM Clustering}
\usage{
neokm(
  x,
  centers,
  alpha = 0.3,
  beta = 0.05,
  nstart = 10,
  trace = FALSE,
  iter.max = 20
)
}
\arguments{
\item{x}{A data matrix.}

\item{centers}{The number of clusters or pre-initialized centers.}

\item{alpha}{A number representing overlap.}

\item{beta}{A number representing non-exhaustiveness.}

\item{nstart}{Number of executions to find the best result.}

\item{trace}{Logical, if TRUE, trace progress of the algorithm.}

\item{iter.max}{Maximum number of iterations allowed.}
}
\value{
A list representing the clustering results.
}
\description{
Cluster data using the NEOKM algorithm.
}
\examples{
neokm(iris[, -5], 3)
neokm(iris[, -5], iris[, -5], 1, 2)
}
