% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okm.R
\name{okm}
\alias{okm}
\title{OKM Clustering}
\usage{
okm(x, centers, iter.max = 10, nstart = 1, trace = FALSE, method = "euclid")
}
\arguments{
\item{x}{A data matrix.}

\item{centers}{A number or matrix representing the number of clusters or
pre-initialized centers.}

\item{iter.max}{Maximum number of iterations allowed.}

\item{nstart}{Number of executions to find the best result.}

\item{trace}{Logical, if TRUE, trace progress of the algorithm.}

\item{method}{A string specifying the distance method ('euclid' or
'manhattan').}
}
\value{
A list representing the clustering results.
}
\description{
Overlapping K-Means clustering algorithm.
}
\examples{
okm(iris[, -5], 3)
}
