% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{plot.interval}
\alias{plot.interval}
\title{Plot Interval Data}
\usage{
\method{plot}{interval}(x, ...)
}
\arguments{
\item{x}{An interval object.}

\item{...}{Additional graphical parameters such as 'col' and 'add'.}
}
\description{
Create a plot representation of interval data.
}
\examples{
plot(iaggregate(iris, 5))
plot(iaggregate(iris, 5), col = 4)
plot(iaggregate(iris, 5), add = TRUE)
}
