% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_icmeans.R
\name{print.icmeans}
\alias{print.icmeans}
\title{Print Method for Fuzzy Icmeans Clustering}
\usage{
\method{print}{icmeans}(x, ...)
}
\arguments{
\item{x}{An icmeans object.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
Print method for displaying results of fuzzy icmeans clustering.
}
