% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ikmeans.R
\name{print.ikmeans}
\alias{print.ikmeans}
\title{Print Method for Ikmeans Clustering}
\usage{
\method{print}{ikmeans}(x, ...)
}
\arguments{
\item{x}{An ikmeans object.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
Print method for displaying results of ikmeans clustering.
}
