% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineokm.R
\name{print.ineokm}
\alias{print.ineokm}
\title{Print Method for Ineokm Clustering}
\usage{
\method{print}{ineokm}(x, ...)
}
\arguments{
\item{x}{An ineokm object.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
Print method for displaying results of ineokm clustering.
}
