% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{print.interval}
\alias{print.interval}
\title{Interval Data Print}
\usage{
\method{print}{interval}(x, ...)
}
\arguments{
\item{x}{An interval object.}

\item{...}{Additional arguments passed to print.}
}
\description{
Custom print method for interval objects.
}
\examples{
print(inter_city)
}
