% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iokm.R
\name{print.iokm}
\alias{print.iokm}
\title{Print Method for IOKM Clustering}
\usage{
\method{print}{iokm}(x, ...)
}
\arguments{
\item{x}{An IOKM object.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
Print method for displaying results of IOKM clustering.
}
