% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okm.R
\name{print.okm}
\alias{print.okm}
\title{Print Method for OKM Clustering}
\usage{
\method{print}{okm}(x, ...)
}
\arguments{
\item{x}{An OKM object.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
Print method for displaying results of OKM clustering.
}
