% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r1okm.R
\name{print.r1okm}
\alias{print.r1okm}
\title{Print Method for R1-OKM Clustering}
\usage{
\method{print}{r1okm}(x, ...)
}
\arguments{
\item{x}{An R1OKM object.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
Custom print method for displaying R1OKM clustering results.
}
