% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2okm.R
\name{print.r2okm}
\alias{print.r2okm}
\title{Print Method for R2-OKM Clustering}
\usage{
\method{print}{r2okm}(x, ...)
}
\arguments{
\item{x}{An R2OKM object.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
Custom print method for displaying R2OKM clustering results.
}
