% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r1okm.R
\name{r1okm}
\alias{r1okm}
\title{R1-OKM Clustering}
\usage{
r1okm(x, centers, alpha = 0, nstart = 10, trace = FALSE, iter.max = 20)
}
\arguments{
\item{x}{A data matrix.}

\item{centers}{The number of clusters or a matrix of initial centers.}

\item{alpha}{A numeric parameter controlling clustering behavior.}

\item{nstart}{Number of executions to find the best result.}

\item{trace}{Logical, if TRUE, provides progress information.}

\item{iter.max}{Maximum number of iterations allowed.}
}
\value{
A list representing the clustering results.
}
\description{
Cluster data using the R1OKM algorithm.
}
\examples{
r1okm(iris[, -5], 3)
r1okm(iris[, -5], 3, alpha = -0.5)
r1okm(iris[, -5], iris[, -5], alpha = 1)
}
