% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2okm.R
\name{r2okm}
\alias{r2okm}
\title{R2-OKM Clustering}
\usage{
r2okm(x, centers, lambda = 0, nstart = 10, trace = FALSE, iter.max = 20)
}
\arguments{
\item{x}{A data matrix.}

\item{centers}{The number of clusters or a matrix of initial centers.}

\item{lambda}{A numeric parameter controlling clustering behavior.}

\item{nstart}{Number of executions to find the best result.}

\item{trace}{Logical, if TRUE, provides progress information.}

\item{iter.max}{Maximum number of iterations allowed.}
}
\value{
A list representing the clustering results.
}
\description{
Cluster data using the R2OKM algorithm.
}
\examples{
r2okm(iris[, -5], 3)
r2okm(iris[, -5], 3, lambda = 0.3)
r2okm(iris[, -5], iris[, -5], lambda = 1)
}
