% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{read.interval}
\alias{read.interval}
\title{Load 3D Interval Array from CSV}
\usage{
read.interval(..., row.names = FALSE, class = NULL)
}
\arguments{
\item{...}{Additional arguments passed to read.csv.}

\item{row.names}{Logical indicating if the first column contains row names.}

\item{class}{The column index of class labels (NULL if not present).}
}
\description{
Load interval data from a CSV file and convert it into a 3D array.
}
