% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{write.interval}
\alias{write.interval}
\title{Save 3D Interval Array to CSV}
\usage{
write.interval(x, ..., class = FALSE)
}
\arguments{
\item{x}{An interval object.}

\item{...}{Additional arguments passed to write.csv.}

\item{class}{Logical indicating whether to write the class column at the end.}
}
\description{
Save interval data to a CSV file.
}
