% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.array.interval}
\alias{as.array.interval}
\title{Converts an interval object to an array representation.}
\usage{
\method{as.array}{interval}(x, ...)
}
\arguments{
\item{x}{An interval object to be converted.}

\item{...}{Additional arguments to be passed to as.array().}
}
\value{
An array representation of the interval.
}
\description{
Converts an interval object to an array representation.
}
\examples{
as.array(inter_city)
}
