% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.array.interval}
\alias{as.array.interval}
\title{Interval to array}
\usage{
\method{as.array}{interval}(x, ...)
}
\arguments{
\item{x}{an interval.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Interval to array
}
\examples{
as.array(inter_city)
}
