% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.data.frame.interval}
\alias{as.data.frame.interval}
\title{Converts an interval object to a data frame representation.}
\usage{
\method{as.data.frame}{interval}(x, ...)
}
\arguments{
\item{x}{An interval object to be converted.}

\item{...}{dditional arguments to be passed to as.data.frame().}
}
\value{
A data frame representation of the interval.
}
\description{
Converts an interval object to a data frame representation.
}
\examples{
as.data.frame(inter_city)
}
