% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.data.frame.interval}
\alias{as.data.frame.interval}
\title{Interval to data.frame}
\usage{
\method{as.data.frame}{interval}(x, ...)
}
\arguments{
\item{x}{an interval.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Interval to data.frame
}
\examples{
as.data.frame(inter_city)
}
