% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.interval.array}
\alias{as.interval.array}
\title{Converts an array to an interval object.}
\usage{
\method{as.interval}{array}(x)
}
\arguments{
\item{x}{An array to be converted to an interval object.}
}
\value{
An interval object constructed from the array if it meets the
requirements, otherwise attempts to convert it to a matrix first.
}
\description{
The array must have three
dimensions, with the second dimension of size 2, representing the minimum
and maximum values.
}
\examples{
as.interval(array(1:12, dim = c(2, 2, 3)))
}
