% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.interval.default}
\alias{as.interval.default}
\title{Provides a default method for converting unsupported data types to interval.}
\usage{
\method{as.interval}{default}(x)
}
\arguments{
\item{x}{An object that does not have a supported conversion method.}
}
\value{
`NULL`, indicating no conversion is possible.
}
\description{
Provides a default method for converting unsupported data types to interval.
}
