% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.interval.matrix}
\alias{as.interval.matrix}
\title{matrix to interval}
\usage{
\method{as.interval}{matrix}(x)
}
\arguments{
\item{x}{an interval.}
}
\description{
matrix to interval
}
\examples{
as.interval(matrix(1:12, 3, 4))
}
