% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.interval.numeric}
\alias{as.interval.numeric}
\title{numeric to interval}
\usage{
\method{as.interval}{numeric}(x)
}
\arguments{
\item{x}{an interval.}
}
\description{
numeric to interval
}
\examples{
as.interval(1:6)
}
