% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.matrix.interval}
\alias{as.matrix.interval}
\title{Interval to matrix}
\usage{
\method{as.matrix}{interval}(x, ...)
}
\arguments{
\item{x}{an interval.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Interval to matrix
}
\examples{
as.matrix(inter_city)
}
