% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.vector.interval}
\alias{as.vector.interval}
\title{Interval to vector}
\usage{
\method{as.vector}{interval}(x, ...)
}
\arguments{
\item{x}{an interval.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Interval to vector
}
\examples{
as.vector(inter_city)
}
