% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.vector.interval}
\alias{as.vector.interval}
\title{Converts an interval object to its vector representation.}
\usage{
\method{as.vector}{interval}(x, ...)
}
\arguments{
\item{x}{An interval object to be converted.}

\item{...}{Additional arguments to be passed to as.vector().}
}
\value{
A numeric vector where each consecutive pair of values represents
an interval.
}
\description{
Converts an interval object to its vector representation.
}
\examples{
as.vector(inter_city)
}
