% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cluster_color}
\alias{cluster_color}
\title{Clustering color for plotting.}
\usage{
cluster_color(x)
}
\arguments{
\item{x}{A cluster.}
}
\description{
Get color from clustering cluster for data plotting.
}
\examples{
plot(iris[,1:2], col=cluster_color(neokm(iris, 2, .2, .05)$cluster))
}
