% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{degree2logical}
\alias{degree2logical}
\title{Degree matrix to logical}
\usage{
degree2logical(x, t = min(apply(x, 1, max)))
}
\arguments{
\item{x}{A matrix of membership degree.}

\item{t}{The threshold to assign over.}
}
\description{
Transform a membership degree matrix into logical matrix by threshold.
}
