% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{iaggregate}
\alias{iaggregate}
\title{Interval aggregation.}
\usage{
iaggregate(data, col = 1)
}
\arguments{
\item{data}{Some data to aggregate.}

\item{col}{A number (the column index to aggregate on).}
}
\description{
Aggregate data to 3D interval array.
}
\examples{
iaggregate(iris, col=5)
iaggregate(rock, col=4)
iaggregate(cars, col=1)
}
