% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{ibind}
\alias{ibind}
\title{Interval binding.}
\usage{
ibind(..., class = FALSE)
}
\arguments{
\item{...}{All intervals to bind.}

\item{class}{set to true to create class by bind interval, false bind class attribute.}
}
\description{
Bind intervals.
}
\examples{
ibind(iaggregate(iris,5), iaggregate(iris,5))
ibind(iaggregate(iris,5), iaggregate(iris,5), iaggregate(iris,5), class=TRUE)
}
