% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{igenerate}
\alias{igenerate}
\title{Interval generate.}
\usage{
igenerate(n, ...)
}
\arguments{
\item{n}{the number of elements to generate.}

\item{...}{Vectors (center mean, center sd, half size mean, half size sd).}
}
\description{
Generate intervals from normal law.
}
\examples{
igenerate(1, c(0,1,2,1))
igenerate(1, c(0,1,2,1), c(100,1,2,1))
}
