% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{is.interval}
\alias{is.interval}
\title{Determines if an object is a strictly valid interval object.}
\usage{
is.interval(x)
}
\arguments{
\item{x}{An R object to be tested.}
}
\value{
A logical value indicating whether the object is a valid interval.
}
\description{
Determines if an object is a strictly valid interval object.
}
\examples{
is.interval(inter_city)
is.interval(1:4)
}
