% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{is.interval}
\alias{is.interval}
\title{Interval data test}
\usage{
is.interval(x)
}
\arguments{
\item{x}{an R object.}
}
\description{
Test if its argument is a (strict) interval.
}
\examples{
is.interval(inter_city)
}
