% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{measure}
\alias{measure}
\title{Visualisation of overlap by seuil.}
\usage{
measure(x, min = 0, max = 1, step = 0.1)
}
\arguments{
\item{x}{A matrix of membership degree.}

\item{min}{The minimum value of degree.}

\item{max}{The maximum value of degree.}

\item{step}{The step of degree.}
}
\description{
Generate plot of overlap by seuil, used to help to defined the best seuil .
}
