% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neokm.R
\name{neokm}
\alias{neokm}
\title{neokm clustering.}
\usage{
neokm(x, centers, alpha = 0.3, beta = 0.05, nstart = 10, trace = FALSE,
  iter.max = 20)
}
\arguments{
\item{x}{An data matrix.}

\item{centers}{A number, number of cluster for clustering or pre init centers.}

\item{alpha}{A number (overlap).}

\item{beta}{A number (non-exhaustiveness).}

\item{nstart}{A number, number of execution to find the best result.}

\item{trace}{A boolean, tracing information on the progress of the algorithm is produced.}

\item{iter.max}{the maximum number of iterations allowed.}
}
\description{
Culster data with neokm algorithm.
}
\examples{
neokm(iris[,-5], 3)
neokm(iris[,-5], iris[,-5], 1, 2)
}
