% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okm.R
\name{okm}
\alias{okm}
\title{OKM clustering.}
\usage{
okm(X, centers, iter.max = 10, nstart = 1, trace = FALSE,
  method = "euclid")
}
\arguments{
\item{X}{A data.}

\item{centers}{A number or matrix, number of cluster for clustering or pre init centers.}

\item{iter.max}{the maximum number of iterations allowed.}

\item{nstart}{A number, number of execution to find the best result.}

\item{trace}{A boolean, tracing information on the progress of the algorithm is produced.}

\item{method}{A string ('euclid': Euclidian distance, 'manhattan': Manhattan distance).}
}
\description{
Overlapping Kmeans algorithm.
}
\examples{
okm(iris[,-5], 3)
}
