% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{plot.interval}
\alias{plot.interval}
\title{Generates a visual representation of interval data as rectangles on a plot.}
\usage{
\method{plot}{interval}(x, ...)
}
\arguments{
\item{x}{An interval object to be plotted.}

\item{...}{Additional graphical parameters such as `col` and `add`.}
}
\value{
No return value, it plot the interval.
}
\description{
Generates a visual representation of interval data as rectangles on a plot.
}
\examples{
plot(iaggregate(iris, 5))
plot(iaggregate(iris, 5), col = 4)
plot(iaggregate(iris, 5), add = TRUE)
}
