% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_icmeans.R
\name{print.icmeans}
\alias{print.icmeans}
\title{Displays the results of fuzzy icmeans clustering in a readable format.}
\usage{
\method{print}{icmeans}(x, ...)
}
\arguments{
\item{x}{An `icmeans` object resulting from the `fuzzy_icmeans` function.}

\item{...}{Additional arguments passed to print().}
}
\value{
No return value, it prints the clustering results to the console.
}
\description{
Displays the results of fuzzy icmeans clustering in a readable format.
}
