% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ikmeans.R
\name{print.ikmeans}
\alias{print.ikmeans}
\title{Displays the results of ikmeans clustering in a readable format.}
\usage{
\method{print}{ikmeans}(x, ...)
}
\arguments{
\item{x}{An `ikmeans` object resulting from the `ikmeans` function.}

\item{...}{Additional arguments passed to print().}
}
\value{
No return value, it prints the clustering results to the console.
}
\description{
Displays the results of ikmeans clustering in a readable format.
}
