% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ikmeans.R
\name{print.ikmeans}
\alias{print.ikmeans}
\title{Ikmeans print}
\usage{
\method{print}{ikmeans}(x, ...)
}
\arguments{
\item{x}{An IKmeans object.}

\item{...}{Other options from print.}
}
\description{
Print override for ikmeans
}
