% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineokm.R
\name{print.ineokm}
\alias{print.ineokm}
\title{Ineokm print}
\usage{
\method{print}{ineokm}(x, ...)
}
\arguments{
\item{x}{An IKmeans object.}

\item{...}{Other options from print.}
}
\description{
Print override for ineokm
}
