% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineokm.R
\name{print.ineokm}
\alias{print.ineokm}
\title{Displays the results of Neo-KM clustering in a user-friendly format.}
\usage{
\method{print}{ineokm}(x, ...)
}
\arguments{
\item{x}{An `ineokm` object resulting from the `ineokm` function.}

\item{...}{Additional arguments passed to print().}
}
\value{
No return value, it prints the clustering results to the console.
}
\description{
Displays the results of Neo-KM clustering in a user-friendly format.
}
