% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{print.interval}
\alias{print.interval}
\title{Interval data print}
\usage{
\method{print}{interval}(x, ...)
}
\arguments{
\item{x}{an interval.}

\item{...}{Other options from print.}
}
\description{
Print override for interval
}
\examples{
print(inter_city)
}
