% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{print.interval}
\alias{print.interval}
\title{Custom print method for displaying interval objects in a readable format.}
\usage{
\method{print}{interval}(x, ...)
}
\arguments{
\item{x}{An interval object to be printed.}

\item{...}{Additional arguments passed to the underlying print() function.}
}
\value{
No return value, it prints the interval to the console.
}
\description{
Custom print method for displaying interval objects in a readable format.
}
\examples{
print(inter_city)
}
