% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iokm.R
\name{print.iokm}
\alias{print.iokm}
\title{Displays the results of IOKM clustering in a user-friendly format.}
\usage{
\method{print}{iokm}(x, ...)
}
\arguments{
\item{x}{An `iokm` object resulting from the `iokm` function.}

\item{...}{Additional arguments passed to print().}
}
\value{
No return value, it prints the clustering results to the console.
}
\description{
Displays the results of IOKM clustering in a user-friendly format.
}
