% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neokm.R
\name{print.neokm}
\alias{print.neokm}
\title{Displays the results of NEOKM clustering in a user-friendly format.}
\usage{
\method{print}{neokm}(x, ...)
}
\arguments{
\item{x}{A `neokm` object resulting from the `neokm` function.}

\item{...}{Additional arguments passed to print().}
}
\value{
No return value, it prints the clustering results to the console.
}
\description{
Displays the results of NEOKM clustering in a user-friendly format.
}
