% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neokm.R
\name{print.neokm}
\alias{print.neokm}
\title{NEOKM print}
\usage{
\method{print}{neokm}(x, ...)
}
\arguments{
\item{x}{An NEOKM object.}

\item{...}{Other options from print.}
}
\description{
Print override for NEOKM
}
