% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okm.R
\name{print.okm}
\alias{print.okm}
\title{Displays the results of OKM clustering in a readable format.}
\usage{
\method{print}{okm}(x, ...)
}
\arguments{
\item{x}{An OKM object resulting from the `okm` function.}

\item{...}{Additional arguments passed to print().}
}
\value{
No return value, it prints the clustering results to the console.
}
\description{
Displays the results of OKM clustering in a readable format.
}
