% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r1okm.R
\name{print.r1okm}
\alias{print.r1okm}
\title{R1-OKM print}
\usage{
\method{print}{r1okm}(x, ...)
}
\arguments{
\item{x}{An NEOKM object.}

\item{...}{Other options from print.}
}
\description{
Print override for R1-OKM
}
