% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r1okm.R
\name{print.r1okm}
\alias{print.r1okm}
\title{Displays the results of R1-OKM clustering in a readable format.}
\usage{
\method{print}{r1okm}(x, ...)
}
\arguments{
\item{x}{An R1-OKM object resulting from the `r1okm` function.}

\item{...}{Additional arguments passed to print().}
}
\value{
No return value, it prints the clustering results to the console.
}
\description{
Displays the results of R1-OKM clustering in a readable format.
}
