% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2okm.R
\name{print.r2okm}
\alias{print.r2okm}
\title{R2-OKM print}
\usage{
\method{print}{r2okm}(x, ...)
}
\arguments{
\item{x}{An NEOKM object.}

\item{...}{Other options from print.}
}
\description{
Print override for R2-OKM
}
