% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2okm.R
\name{print.r2okm}
\alias{print.r2okm}
\title{Displays the results of R2-OKM clustering in a readable format.}
\usage{
\method{print}{r2okm}(x, ...)
}
\arguments{
\item{x}{An R2-OKM object resulting from the `r2okm` function.}

\item{...}{Additional arguments passed to print().}
}
\value{
No return value, it prints the clustering results to the console.
}
\description{
Displays the results of R2-OKM clustering in a readable format.
}
