% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2okm.R
\name{r2okm}
\alias{r2okm}
\title{r2okm clustering.}
\usage{
r2okm(x, centers, lambda = 0, nstart = 10, trace = FALSE, iter.max = 20)
}
\arguments{
\item{x}{An data matrix.}

\item{centers}{A number, number of cluster for clustering or pre init centers.}

\item{lambda}{A number.}

\item{nstart}{A number, number of execution to find the best result.}

\item{trace}{A boolean, tracing information on the progress of the algorithm is produced.}

\item{iter.max}{the maximum number of iterations allowed.}
}
\description{
Culster data with r2okm algorithm.
}
\examples{
r2okm(iris[,-5], 3)
r2okm(iris[,-5], 3, .3)
r2okm(iris[,-5], iris[,-5], 1)
}
