% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{read.interval}
\alias{read.interval}
\title{Interval loading.}
\usage{
read.interval(..., row.names = FALSE, class = NULL)
}
\arguments{
\item{...}{Other options from read.}

\item{row.names}{Row have column names ?}

\item{class}{The column of class (empty if not).}
}
\description{
Load a 3D interval array from csv.
}
