% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{read.interval}
\alias{read.interval}
\title{Reads a CSV file and converts the data into a 3D interval array.}
\usage{
read.interval(..., row.names = FALSE, class = NULL)
}
\arguments{
\item{...}{Additional arguments passed to read.csv().}

\item{row.names}{Logical indicating if the first column contains row names.}

\item{class}{The column index of class labels (set to `NULL` if not present).}
}
\value{
A structured interval object representing the data from the CSV file.
}
\description{
Reads a CSV file and converts the data into a 3D interval array.
}
