% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{write.interval}
\alias{write.interval}
\title{Writes an interval object to a CSV file.}
\usage{
write.interval(x, ..., class = FALSE)
}
\arguments{
\item{x}{An interval object to be saved.}

\item{...}{Additional arguments passed to write.csv().}

\item{class}{Logical indicating whether to add the class column in the CSV.}
}
\value{
No return value, it saves the interval to the given CSV file.
}
\description{
Writes an interval object to a CSV file.
}
