% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{write.interval}
\alias{write.interval}
\title{Interval saving.}
\usage{
write.interval(x, ..., class = FALSE)
}
\arguments{
\item{x}{An interval.}

\item{...}{Other options from write.}

\item{class}{Write the class column at the end ?.}
}
\description{
Save 3D interval array to csv to open later with iload.
}
